Changed avgRating to INTEGER for testing

+++ backend/db/migrations/20240213050038-create-spot.js
@@ -60,7 +60,7 @@ module.exports = {
           allowNull: false,
         },
         avgRating: {
-          type: Sequelize.DECIMAL,
+          type: Sequelize.INTEGER,
         },
         previewImage: {
           type: Sequelize.STRING,
